+++ lib/flux-core/src/slice/mod.rs
@@ -101,10 +101,10 @@ impl<T> [T] {
     fn split_at_mut_checked(&mut self, mid: usize) -> Option<(&mut [T], &mut [T])>;
 
     #[no_panic]
-    #[spec(fn(&Self[@n]) -> *const{p: ptr_size(p) == n * T::size_of() && ptr_offset(p) == 0} T)]
+    #[spec(fn(&Self[@n]) -> *const{p: ptr_size(p) == n} T)]
     fn as_ptr(&self) -> *const T;
 
     #[no_panic]
-    #[spec(fn(&mut Self[@n]) -> *mut{p: ptr_size(p) == n * T::size_of() && ptr_offset(p) == 0} T)]
+    #[spec(fn(&mut Self[@n]) -> *mut{p: ptr_size(p) == n} T)]
     fn as_mut_ptr(&mut self) -> *mut T;
 }
