+++ tests/tests/neg/surface/raw_ptr01.rs
@@ -1,0 +1,37 @@
+extern crate flux_core;
+
+#[flux::spec(fn (ptr: *const{v: ptr_size(v) > 1} i32))]
+pub fn test_add_ex(ptr: *const i32) {
+    unsafe {
+        let _val0 = std::ptr::read(ptr.add(0));
+        let _val1 = std::ptr::read(ptr.add(1));
+        let _val2 = std::ptr::read(ptr.add(2)); //~ ERROR: refinement type error
+    }
+}
+
+#[flux::spec(fn (ptr: *const{v: ptr_size(v) == 2} i32))]
+pub fn test_add_ix(ptr: *const i32) {
+    unsafe {
+        let _val0 = std::ptr::read(ptr.add(0));
+        let _val1 = std::ptr::read(ptr.add(1));
+        let _val2 = std::ptr::read(ptr.add(2)); //~ ERROR: refinement type error
+    }
+}
+
+#[flux::spec(fn (ptr: *mut{v: ptr_size(v) > 1} i32))]
+pub fn test_add_mut_ex(ptr: *mut i32) {
+    unsafe {
+        std::ptr::write(ptr.add(0), 10);
+        std::ptr::write(ptr.add(1), 20);
+        std::ptr::write(ptr.add(2), 30); //~ ERROR: refinement type error
+    }
+}
+
+#[flux::spec(fn (ptr: *mut{v: ptr_size(v) == 2} i32))]
+pub fn test_add_mut_ix(ptr: *mut i32) {
+    unsafe {
+        std::ptr::write(ptr.add(0), 10);
+        std::ptr::write(ptr.add(1), 20);
+        std::ptr::write(ptr.add(2), 30); //~ ERROR: refinement type error
+    }
+}
