+++ tests/tests/neg/surface/raw_ptr02.rs
